CourtMaster/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-multipart
alembic
python-jose[cryptography]
passlib[bcrypt]
openpyxl
